move localstorage service
